small fixes: capitalizing links
now that the i was changed to a = in the html part, the murmuration description should work. gonna test. the "murmuration" was 
changed a "Murmuration", as according to page source the m was actually capital; this should fix the issue of the code working for the mrhs section but not the other two. 
space between < and p for the hostingmeet description; will fix.